+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.0.1
+* Code changes to use more the postcss API instead of concatenating string
+* New methods have been added to remove the directive comments from the output
+
 ## 1.0.0
 * First commit